requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.3
python-socketio==5.8.0
python-engineio==4.4.1
Pillow==9.5.0
uiautomator2==2.16.17
adbutils==1.2.15
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
colorama==0.4.6
eventlet==0.33.3
dnspython==2.3.0
gevent==22.10.2
python-dotenv==1.0.0
six==1.16.0
websocket-client==1.5.1 